Fixes #736 - Added further tests.  Implemented forwarding for mobile addresses.
